backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-socketio==5.3.4
python-socketio==5.9.0
eventlet==0.33.3
RPi.GPIO==0.7.1
